Split classes into separate files in directory
